Disable code coverage and Werror to build with clang
Summary:
In autoconf, clang passes the `test $GCC = yes` test. This seems surprising but it is actually expected, because the way autoconf tests this is by checking the `__GNU_C__` macro. And clang does have it since it supports the GNU C dialect. In this case, this behavior has caused some issues. See the changes listed below.
* Pass `--disable-coverage` to autoconf. This enables instrumentation profiling (via `-fprofile-arcs`), which in clang's case requires the compiler-rt project to be built and linked. This can be cumbersome as we don't normally build that project.
* Remove `-Werror`. In the newest clang 15, more diagnostics have been added and it reports more warnings than the system-default GCC. So the installation fails. Since we don't change the benchmark that much, it makes sense to remove this flag instead of fixing in the source code. (fixing in source code requires patching more changes which makes it inconvenient).

Reviewed By: excelle08

Differential Revision: D37836964

fbshipit-source-id: 49a2f59c6daf4b170c7f1b5d1e84764dffb94eac