Update formula for cpu cores range calculation (#36)
Summary:
For TaoBench2x, with current core range calculation, one tao_server instance is pinned to physical cpu cores and the other tao_server instance is pinned to vcpu cores. As a result, the QPS from each instance is not balanced.

Core pinning in a way that the physical cpu cores and their corresponding vcpu get bound to one instance of tao_server will reduce ccx traffic / latency.

In my experiments, this method gave an uplift of 5-7% in the total QPS for TaoBench2x.

Pull Request resolved: https://github.com/facebookresearch/DCPerf/pull/36

Reviewed By: adhanotia

Differential Revision: D46729463

Pulled By: excelle08

fbshipit-source-id: e0f47686bb4ec63f5a69610f8bf11bcb172d38d3