add arm-specific job for feedsim
Summary:
As reported by arm and investigated by us, icache buster has a different behavior on arm vs on x86, leading to extra hotness of the function. This diff add an arm-specific job for feedsim that lower the number of icache buster iterations, so that the hotness matches.

didn't choose to implement it via `extra_args` because it could be overwritten by user specified other parameters (like `-q` in the readme example).

Note that this is a temp and quick fix. an icache buster replacement can better resolve it (I believe it's on the plan already)

Reviewed By: excelle08

Differential Revision: D68851663

fbshipit-source-id: 6d1b1c42b62bcf853d403d9a009039c61ab7cac3