Revert the portion of D70239810 that changed DCPerf codebase
Summary: D70239810 authored by a bot changed the codebase of DCPerf (fbcode//cea/chips/benchpress) to use some modern features that are available in Python 3.10+. However this will break DCPerf's OSS run, considering the default Python interpreter in CentOS 9 is 3.9 so some of the newer features are not supported. Therefore we need to revert the changes.

Differential Revision: D70342397