Revert the portion of D70239810 that changed DCPerf codebase (#79)
Summary:
Pull Request resolved: https://github.com/facebookresearch/DCPerf/pull/79

D70239810 authored by a bot changed the codebase of DCPerf (fbcode//cea/chips/benchpress) to use some modern features that are available in Python 3.10+. However this will break DCPerf's OSS run, considering the default Python interpreter in CentOS 9 is 3.9 so some of the newer features are not supported. Therefore we need to revert the changes.

Reviewed By: YifanYuan3

Differential Revision: D70342397

fbshipit-source-id: 1d1b4926daa1240dc15787a654c57a420550e76e