Fix frontend/backend boundness counter on Grace
Summary:
This was previously merged to main branch: https://github.com/facebookresearch/DCPerf/pull/104

Do we still need to commit it to the v2 branch? Or should we resolve such conflicts later?

I'm not familiar enough with the current workflow of DCPerf, so just creating the diff here. Please let me know whether I should ship or abandon it. Thanks!

Reviewed By: excelle08

Differential Revision: D74264455