Enabling Django to take a snapshot and restore from it (#126)
Summary:
Pull Request resolved: https://github.com/facebookresearch/DCPerf/pull/126

The Django benchmark generates data and load dataset at the benchmarking time. This diff mainly enables Django to (i) take a snapshot, (ii) use a snapshot to load a dataset. In addition this diff, creates a new job for Django where the duration is smaller and uses a snapshot.

Unfortunately, the Cassandra version that we have in the DCPerf is very old, and taking a snapshot and restore from it was different from the documentation that I found for V3 and I only managed to make this work by trying and error of many things.

So, the next step for this diff might be updating the Cassandar for DCPerf

Reviewed By: YifanYuan3

Differential Revision: D75101136