Add adsim as a benchmark in DCPerf
Summary:
This update introduces a new end-to-end benchmark, `adsim`, to the AI benchmark suite. Adsim is a configurable benchmark designed for AdServing Services. It simulates the behavior of production workloads by executing several computational tasks (kernels) based on a configuration file. The primary goal of this benchmark is to provide an emulation of end-to-end AI inference workloads in the production. As the first step, this diff gives all the build recipe and necessary change to ship the adsim first into the DCPerf. Next step to find the configuration that can representatively simulate the production AI inference workload.

#### How to Run the New Benchmark

To install and execute the `adsim` benchmark, use the following commands:

```bash
./benchpress_cli.py -b ai install adsim  # Install the adsim benchmark

./benchpress_cli.py -b ai run adsim      # Run the adsim benchmark
```

#### FILES ADDED AND MODIFIED
```
- benchpress
  - configs
    - benchmarks_ai.yml  // Added the new adsim benchmark.
    - jobs_ai.yml        // Added new adsim benchmark jobs.
- packages
  - adsim
    - src/               // Source code for adsim linked to the adsim codebase.
    - patches/           // Patches required to build adsim.
    - buildfiles         // CMake build files.
    - configs            // Configuration file for running the adsim server.
    - install_adsim.sh   // New installation script for the adsim benchmark.
    - cleanup_deser.sh   // New cleanup script for the adsim benchmark.
    - build-deps.sh      // Builds all dependencies needed for adsim.
    - build-adsim.sh     // Builds the adsim.
    - config.sh          // Defines globally used variables.
    - install_fbgemm.sh  // Script to install fbgemm.
    - qbs_search.sh      // Script to search for the correct QPS.
    - run_adsim.py       // Script to run the client and server.
    - adsim_config.py    // Sets up the configuration for the server.
- BUCK                   // Updated to include the new benchmark components.
...
```

Reviewed By: excelle08

Differential Revision: D78776869