Improve finally clause when benchmark/job/toolchain spec given as argument doesn't exist
Summary: When inadvertently providing a bad benchmark/job/toolchain spec name, the logger error message correctly says it doesn't exist, but it's buried behind the other exception that occurs in the "finally" clause. This just adds a fix so we don't throw an exception, so the original error is easier to see.

Differential Revision: D83190156