Feedsim mini : recoridng benchmark timestamps to be used for metric filtering
Summary:
This diff introduces runtime breakdown tracking capabilities for the DCPerf mini benchmarking infrastructure. It adds utilities to track and log the timing of benchmark operations with process ID (PID) and timestamp information.

This is essential for mini jobs, where the execution time is only a few seconds and we need to filter metrics for the actual benchmark time.

Reviewed By: excelle08

Differential Revision: D84876459