Add auto-scaled benchmark for chm
By default, the worker thread count for the CHM benchmark is set to 80.
This patch introduces an auto-scaled benchmark configuration that sets
the thread count to the number of all cpu threads. User can use
`numactl` or `taskset` to limit the cpus to use, and this benchmark will
use the specified cpu number as the worker threads count.