add scoring mechanism to WDL
Summary:
This diff adds 
(1) baseline perf numbers of the prod-focused WDL microbenchmarks 
(2) script to calculate the scores of the wdl run, based on (1) and (for some microbenchmarks) the weights of different operations in real prod.
(3) updated README for WDL so to provide guidance of best practice of WDL usage.


right now the scoring of `sleef` and `memcmp` is TODO, which will be added in next diff

Reviewed By: mcfi

Differential Revision: D88109786