Replace threads time with benchmark time
The metric used for chm is Mops/sec, the time used in computing is the
total time of every thread's running time, in order to better reflect
the throughput per second, it should be changed to the time of the
benchmark running. In this way, if there is more threads used, the
throughput will be bigger.